requirements.txt
# requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
sqlalchemy==2.0.36
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.0.3
openai==0.28.0
fastapi-mail==1.4.1
psycopg2==2.9.10